tree kem kat externalized
