+++ kat_generation/src/lib.rs
@@ -1,1 +1,1 @@
-mod kat_tree_kem;
+pub mod kat_tree_kem;
